re-organized body code
