+++ fitnick/body/bodyfat/bodyfat.py
@@ -1,50 +1,11 @@
+from sqlalchemy.dialects.postgresql import insert
 from sqlalchemy.exc import IntegrityError
-from tqdm import tqdm
-
-from fitnick.base.base import get_authorized_client
 from sqlalchemy.orm import sessionmaker
 
-from fitnick.database.database import Database
-from fitnick.time_series import TimeSeries, set_dates
+from fitnick.base.base import get_authorized_client
 from fitnick.body.models.bodyfat import BodyFatRecord, bodyfat_table
-from fitnick.body.models.weight import WeightRecord, weight_table
-
-from sqlalchemy.dialects.postgresql import insert
-
-
-class WeightTimeSeries(TimeSeries):
-    def __init__(self, config):
-        super().__init__(config)
-        self.config['schema'] = 'weight'
-        self.config['resource'] = 'weight'
-        return
-
-    @staticmethod
-    def parse_response(data):
-        rows = []
-        for record in data['body-weight']:
-            row = WeightRecord(
-                date=record['dateTime'],
-                pounds=record['value'])
-            rows.append(row)
-
-        return rows
-
-    def insert_data(self):
-        data = self.query()
-        parsed_rows = self.parse_response(data)
-        db = Database(self.config['database'], schema=self.config['schema'])
-        session = sessionmaker(bind=db.engine)()
-        for row in parsed_rows:
-            insert_statement = insert(weight_table).values(
-                date=row.date,
-                pounds=row.pounds
-            )
-            session.execute(insert_statement)
-            session.commit()
-
-        session.close()
-        return
+from fitnick.database.database import Database
+from fitnick.time_series import set_dates
 
 
 class BodyFat:
@@ -109,7 +70,7 @@ class BodyFat:
             except IntegrityError:  # record already exists
                 pass
 
-        session.close() 
+        session.close()
 
     def get_bodyfat_rows(self):
         response = self.query()
